### 2018-05-14 #217 @pachamaltese
oec package
oec provides an easy way to obtain data from the Observatory of Economic Complexity by accessing its API.

```
Package: oec
Title: Observatory of Economic Complexity API Wrapper and Utility Program
Version: 2.8.2
  person("Mauricio", "Vargas S.", role = c("aut", "cre", "cph"), email = "oec@media.mit.edu"),
  person("Manuel", "Aristaran", role = c("ctb")),
  person("Pablo", "Paladino", role = c("ctb")),
  person("Gabriela", "Perez", role = c("ctb")),
  person(family = "UN Comtrade", role = c("dtc")),
  person(family = "MIT Media Lab", role = c("dtc")),
  person(family = "Datawheel", role = c("fnd"))
  )
Description: Access The Observatory of Economic Complexity API from R to download international trade data.
  curl,
  jsonlite
    covr
```

https://github.com/pachamaltese/oec-r-package

databases, because the package connects to an API and does 3 or more API calls to simplify things for the final user who wants imports/exports and some metrics such as % of increase/decrease.

Non-expert users that use international trade data. This can also be targeted to intermediate/advanced users who can benefit from the speed and short syntax that this package provides.

Not at the moment (in 2 yrs this is the only one)

No.

None.

No.

I don't really know.
