### 2018-05-15 #217 @maelle
Also, could you please expand here on the added value of the data processing offered by your package as opposed to using the original data sources?
