### 2018-05-15 #217 @pachamaltese
@maelle thanks, no worries, let me check my laptop hdd vs my external drive, I probably forgot to move something

About the added value if I run this command `getdata_batch("chl", "chn", 2010, 2015)` I will obtain a tibble with 5 years of data for chilean exports to china but with these additions:

* country codes corrected and written as ISO-3
* full product names
* full country names
* trade balance
* economic complexity measures and ranks
* top importers and exporters
* a color column to use the same palette as atlas.media.mit.edu

On the other hand, with API calls I would need to read https://atlas.media.mit.edu/hs92/export/2010/chl/chn/show/
https://atlas.media.mit.edu/hs92/export/2011/chl/chn/show/
... etc ...
https://atlas.media.mit.edu/hs92/export/2015/chl/chn/show/

and also to read this
https://atlas.media.mit.edu/hs92/export/2010/chl/all/show/
https://atlas.media.mit.edu/hs92/export/2011/chl/all/show/
... etc ...
https://atlas.media.mit.edu/hs92/export/2015/chl/all/show/

and also this
https://atlas.media.mit.edu/hs92/export/2010/all/all/show/
https://atlas.media.mit.edu/hs92/export/2011/all/all/show/
... etc ...
https://atlas.media.mit.edu/hs92/export/2015/all/all/show/

After reading those 15 json files the user would need some additional operations to see the same result as `getdata_batch("chl", "chn", 2010, 2015)`:
* a `left_join` to match API product code (e.g. `2709`) to product name (e.g. petroleum), a mayor drawback here is that I the code-name can't be accessed by using the API
* another `left_join` to match country codes (e.g. `sachl`) to country names (e.g. Chile) with the big problem that the API returns non-ISO codes (e.g. `sachl` vs `chl`), again the user would need a code-name table and would need to crop the strings
* join everything and finally use `purrr` to create a unique tibble
