### 2018-05-19 #217 @pachamaltese
Hi @maelle.
I did put all the pieces together: external disks, google drive, etc. I have pushed a new version with lots of changes as well as a testing unit.
Have a good weekend!
