### 2018-05-24 #217 @pachamaltese
thanks a lot @maelle

happy travelling!!

I have +5 external hdd since my macbook fried itself and I lost tons of tidy useful data :(
I'm still recovering from that accident ;)
