### 2018-05-24 #217 @maelle
### Editor checks:

- [ x] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

Thanks again for your submission @pachamaltese!

Here are a some points that should be solved/discussed before I start looking for reviewers. Note that now that you have everything on GitHub, you shouldn't loose pieces again, which is good news. Sorry to hear about your Macbook! 😢

This process is an ongoing discussion so ask any question/make any remark you might have about my comments!

#### Code

* You use `assign` in functions instead of returning output. Please return output, you don't need to use `assign` here.

* Why do you use `rm` inside a function?

* There is a separate file containing imports (confusingly called exports :wink:). It'd be easier to see where functions are imported from if you used the syntax `jsonlite::fromJSON`. You could then remove the separate R code with imports. Cf http://r-pkgs.had.co.nz/namespace.html#imports

* Your package depends on `purrr` for a single function. Consider replacing it with `do.call` and `rbind`.

* Regarding all the `stopifnot`, have informative error messages when the user entered a wrong parameter value.

* You could have a single exported function instead of two if I follow correctly? It'd take arguments like the batch one, that could be vector of length 1 (the user would input say one or several country pairs). The other function would be an utility function, that wouldn't need to be exported. It might simplify the interface?

* If there is only one function call in a pipeline it might be clearer to remove the pipe.

```r
  origin_destination <- origin_destination %>%
    mutate(trade_exchange_val =
!!sym("export_val") + !!sym("import_val"))

```

would become

```r
  origin_destination <-  mutate(origin_destination,
    trade_exchange_val = !!sym("export_val") + !!sym("import_val"))
```

* In case you might ignore it, and in case it could be useful, there's a [`countrycode`](https://cran.r-project.org/web/packages/countrycode/index.html) package.

#### Data

* Where do the data files in data come from? Why are they pre-prepared, are you sure they won't be updated?

* You could keep their data processing in data-raw.

#### Documentation

* Please add the URL to the data source.

* Are you allowed to use the logo in your README?

* A flowchart of what processing your package does might be useful to explain the added value of using the package and to explain what has been preprocessed or not (what's being downloaded vs what's in data/).

* In the README when mentioning the vignette have a link to the `pkgdown` website.

* Please add @return fields to the documentation of the functions.

* A vignette with an actual use case, and visualizations, would be nice (or an extension of the existing vignette).

* Please explain what the data is e.g. SITC or add an external link.

* Because some chunks of the vignette are unevaluated without a pre-generated output, I can't understand what the "problem" is. You could pre-generate the output.

#### Misc

* Could you rename the repo "oec" like the package?

* The "Maintainer" field in DESCRIPTION is useless because it'll be generated automatically from Authors.

* What are the files .Rapp.history in data/ and .build.timestamp in vignettes/?

* Please add a code coverage integration and badge (via e.g. `usethis::usethis::use_coverage()`)

####  goodpractice output

```r
 ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite
    often. A build date will be added to the package when you perform `R CMD build` on
    it.
  ✖ avoid long code lines, it is bad for readability. Also, many people
    prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R\getdata.R:56:1
    R\getdata_batch.R:44:1
    R\getdata_batch.R:68:1
    tests\testthat\test-oec.R:3:1
    tests\testthat\test-oec.R:15:1
    ... and 10 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
    1:NCOL(...) expressions. They are error prone and result 1:0 if the expression on
    the right hand side is zero. Use seq_len() or seq_along() instead.

    R\getdata_batch.R:66:13
```

---

Reviewers:
Due date:
