### 2018-06-09 #217 @pachamaltese
@maelle it's ok now, I did all the steps from your feedback and just a few are explained why yes/no to it is better for this particular API.
