### 2018-06-09 #217 @pachamaltese
## Maelle' s initial feedback

- [x] You use assign in functions instead of returning output. Please return output, you don't need to use assign here.
- [x] Why do you use rm inside a function? *fixed, it was a bad try to make lighter function calls*
- [x] There is a separate file containing imports (confusingly called exports wink). It'd be easier to see where functions are imported from if you used the syntax jsonlite::fromJSON. You could then remove the separate R code with imports. http://r-pkgs.had.co.nz/namespace.html#imports
- [x] Your package depends on purrr for a single function. Consider replacing it with do.call and rbind. *To (over)simplify now I also use purrr's flatten df function*
- [x] Regarding all the stopifnot, have informative error messages when the user entered a wrong parameter value.
- [x] You could have a single exported function instead of two if I follow correctly? It'd take arguments like the batch one, that could be vector of length 1 (the user would input say one or several country pairs). The other function would be an utility function, that wouldn't need to be exported. It might simplify the interface?
- [x]  If there is only one function call in a pipeline it might be clearer to remove the pipe.

```
origin_destination <- origin_destination %>%
    mutate(trade_exchange_val =
!!sym("export_val") + !!sym("import_val"))
```

would become

```
origin_destination <-  mutate(origin_destination,
    trade_exchange_val = !!sym("export_val") + !!sym("import_val"))
```

- [x] In case you might ignore it, and in case it could be useful, there's a countrycode package. *the OEC uses an ISO modification, so I'm not so sure about it*

### Data

- [x] Where do the data files in data come from? Why are they pre-prepared, are you sure they won't be updated? *everything in data/ comes from atlas.media.mit.edu and corresponds to unaltered trade codes (i.e HS96 was created not to touch HS92 created in 1992, HS02 was an update to HS96, etc so
these tables do not change*
- [x] You could keep their data processing in data-raw. *ok, I'll add a vignette to explain the changes*

- [x] Please add the URL to the data source.
- [x] Are you allowed to use the logo in your README? *yes, I' m a part of the original team*
- [x] A flowchart of what processing your package does might be useful to explain the added value of using the package and to explain what has been preprocessed or not (what's being downloaded vs what's in data/).
- [x] In the README when mentioning the vignette have a link to the pkgdown website.
- [x] Please add @return fields to the documentation of the functions.
- [x] A vignette with an actual use case, and visualizations, would be nice (or an extension of the existing vignette).
- [x] Please explain what the data is e.g. SITC or add an external link.
- [x] Because some chunks of the vignette are unevaluated without a pre-generated output, I can't understand what the "problem" is. You could pre-generate the output.

### Misc

- [x] Could you rename the repo "oec" like the package? *travis also renamed*
- [x] The "Maintainer" field in DESCRIPTION is useless because it'll be generated automatically from Authors.
- [x] What are the files .Rapp.history in data/ and .build.timestamp in vignettes/? *temp files from R CMD, ok now*
- [x] Please add a code coverage integration and badge (via e.g. usethis::usethis::use_coverage())
