### 2018-06-10 #217 @maelle
Thanks a lot @pachamaltese! 😺  I started looking and two small comments already:

* The URL to the `pkgdown` site should be updated now that the repo has a new name.

* You don't export the batch function any more but it's still present in the website docs at least cf http://pacha.hk/oec/reference/index.html
