### 2018-06-10 #217 @maelle
The link to the logo is also broken at the moment.
