### 2018-06-10 #217 @pachamaltese
Great, thanks ! I'll do that before 14.00
And sorry I didn't notice I broke links when renaming

On Sun, Jun 10, 2018, 8:26 AM Maëlle Salmon <notifications@github.com>
wrote:

> Thanks again @pachamaltese <https://github.com/pachamaltese>!
>
> A few more points before I start looking for reviewers
>
>    -
>
>    The URL in the repo description should be updated.
>    -
>
>    The documentation website has to be updated (it still documents the
>    function that was removed)
>    -
>
>    The logo doesn't appear any more in the README.
>    -
>
>    You can now remove my feedback from the README 😉 and have a more
>    prominent link to the pkgdown website in it.
>    -
>
>    You can now add a peer-review badge to your README
>
>
> It'll indicate the package is under review, until after onboarding when
> it'll then indicate it was peer-reviewed.
>
>    -
>
>    Please add a code of conduct to your repo, e.g. via
>    usethis::use_code_of_conduct()
>    -
>
>    Rbuildignore data-raw e.g. via usethis::use_build_ignore("data-raw")
>    -
>
>    The codecov badge has a wrong URL.
>    -
>
>    goodpractice output
>
> It is good practice to
>
>   ✖ write unit tests for all functions, and all package code in
>     general. 89% of code lines are covered by test cases.
>
>     R/getdata.R:44:NA
>     R/getdata.R:47:NA
>     R/getdata.R:50:NA
>     R/getdata.R:55:NA
>     R/getdata.R:59:NA
>     ... and 13 more lines
>
>   ✖ omit "Date" in DESCRIPTION. It is not required and it gets
>     invalid quite often. A build date will be added to the package when you
>     perform `R CMD build` on it.
>   ✖ avoid long code lines, it is bad for readability. Also, many
>     people prefer editor windows that are about 80 characters wide. Try make
>     your lines shorter than 80 characters
>
>     R\getdata.R:47:1
>     R\getdata.R:50:1
>     R\getdata.R:59:1
>     R\getdata.R:64:1
>     R\getdata.R:69:1
>     ... and 5 more lines
>
>   ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
>     1:NCOL(...) expressions. They are error prone and result 1:0 if the
>     expression on the right hand side is zero. Use seq_len() or seq_along()
>     instead.
>
>     R\getdata.R:111:32
>     R\getdata.R:174:26
>     R\getdata.R:200:25
>
> I think that you could easily increase code coverage, looking at
> https://codecov.io/gh/pachamaltese/oec/src/master/R/getdata.R you should
>
>    -
>
>    test that the function returns errors when it should
>    -
>
>    test more use cases (e.g. when destination == "all")
>
> Thanks for all your work!
>
