### 2018-06-10 #217 @maelle
Note that only one vignette appeared on the pkgdown website last time I looked
