### 2018-06-10 #217 @pachamaltese
@maelle ready, I used Jim Hester's `lintr` package for this 2nd round of feedback. Thanks a lot for the useful comments !!
