### 2018-06-13 #217 @maelle
Thanks @aedobbyn and @cimentadaj for accepting to review this package! 😸

Your reviews are due on 2018-07-04. As a reminder here are links to our [bookdown](https://ropenscilabs.github.io/dev_guide/) in particular the [packaging guidelines](https://ropenscilabs.github.io/dev_guide/building.html) and [reviewer's guide](https://ropenscilabs.github.io/dev_guide/onboarding-guide-for-reviewers.html) and to the [review template](https://raw.githubusercontent.com/ropensci/onboarding/master/reviewer_template.md).
