### 2018-06-18 #217 @aedobbyn
Thanks @maelle and good to meet you @pachamaltese! This package looks very useful (and I love the hex sticker!)

I want to make sure I've got the right version of the package. I installed the development version but running any of the `getdata` examples produces the same error in what looks to be `jsonlite:::parse_con`.

Let me know if this is fixed in another branch and I'll re-install from there. My session info is in the reprex below.

```
library(oec)

getdata("chl", "chn", 2015)
#> Valid input. Processing 2015 data...
#> Error in parse_con(txt, bigint_as_char): parse error: premature EOF
#>
#>                      (right here) ------^

devtools::session_info()
#> Session info -------------------------------------------------------------
#>  setting  value
#>  version  R version 3.4.4 (2018-03-15)
#>  system   x86_64, darwin15.6.0
#>  ui       X11
#>  language (EN)
#>  collate  en_US.UTF-8
#>  tz       America/Chicago
#>  date     2018-06-18
#> Packages -----------------------------------------------------------------
#>  package    * version    date
#>  assertthat   0.2.0      2017-04-11
#>  backports    1.1.2      2017-12-13
#>  base       * 3.4.4      2018-03-15
#>  bindr        0.1.1      2018-03-13
#>  bindrcpp     0.2.2      2018-03-29
#>  compiler     3.4.4      2018-03-15
#>  curl         3.2        2018-03-28
#>  datasets   * 3.4.4      2018-03-15
#>  devtools     1.13.5     2018-02-18
#>  digest       0.6.15     2018-01-28
#>  dplyr        0.7.5      2018-05-19
#>  evaluate     0.10.1     2017-06-24
#>  glue         1.2.0.9000 2018-06-18
#>  graphics   * 3.4.4      2018-03-15
#>  grDevices  * 3.4.4      2018-03-15
#>  htmltools    0.3.6      2017-04-28
#>  jsonlite     1.5        2017-06-01
#>  knitr        1.20       2018-02-20
#>  magrittr     1.5        2014-11-22
#>  memoise      1.1.0      2017-04-21
#>  methods    * 3.4.4      2018-03-15
#>  oec        * 2.8.1      2018-06-18
#>  pillar       1.2.1      2018-02-27
#>  pkgconfig    2.0.1      2017-03-21
#>  purrr        0.2.5      2018-05-29
#>  R6           2.2.2      2017-06-17
#>  Rcpp         0.12.17    2018-05-18
#>  rlang        0.2.1      2018-05-30
#>  rmarkdown    1.9        2018-03-01
#>  rprojroot    1.3-2      2018-01-03
#>  stats      * 3.4.4      2018-03-15
#>  stringi      1.2.2      2018-05-02
#>  stringr      1.3.1      2018-05-10
#>  tibble       1.4.2      2018-01-22
#>  tidyselect   0.2.4      2018-02-26
#>  tools        3.4.4      2018-03-15
#>  utils      * 3.4.4      2018-03-15
#>  withr        2.1.2      2018-03-15
#>  yaml         2.1.18     2018-03-08
#>  source
#>  CRAN (R 3.4.0)
#>  CRAN (R 3.4.3)
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.3)
#>  cran (@0.7.5)
#>  CRAN (R 3.4.1)
#>  Github (tidyverse/glue@a2c0f8b)
#>  local
#>  local
#>  CRAN (R 3.4.0)
#>  CRAN (R 3.4.0)
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.0)
#>  CRAN (R 3.4.0)
#>  local
#>  Github (pachamaltese/oec-r-package@00a367b)
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.0)
#>  cran (@0.2.5)
#>  CRAN (R 3.4.0)
#>  cran (@0.12.17)
#>  cran (@0.2.1)
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.3)
#>  local
#>  cran (@1.2.2)
#>  cran (@1.3.1)
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.3)
#>  local
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
```
