### 2018-06-18 #217 @pachamaltese
@aedobbyn Hi. Nice to meet u too! Thanks for letting me know. I installed R on another machine to try to reproduce this and ... it looks as the OEC has a problem

I should explore this, the problem can me an MIT server problem at the moment, but this is problem as the package just verifies internet connection before trying to get data and does not inspect additional things
