### 2018-06-18 #217 @pachamaltese
@aedobbyn *update*: somebody within the OEC team screwed something and not all origins/destinations are available at the moment... the US data is working ok but no cigar trying to compute RCAs that need world flows :(

I shall update when this is fixed

![screenshot from 2018-06-18 13-50-42](https://user-images.githubusercontent.com/10091065/41553077-b5d1e078-72fe-11e8-94c1-d50415011032.png)
