### 2018-06-18 #217 @aedobbyn
Great, thanks for the quick update! Good opportunity to provide informative error messages to the user when the server is down or, in the case of the URL above, nothing is returned at all (I don't even get a 500 in the Chrome console). I'll look out for another update from you!
