### 2018-06-18 #217 @maelle
To test (in unit tests) how the package react to http errors you can use @sckott's `webmockr` package https://github.com/ropensci/webmockr
