### 2018-06-19 #217 @pachamaltese
@maelle: Thanks a again !! at this point you have contributed more to this package than the actual listed ctbs
@aedobbyn: Thanks, now it's fixed. This was a server side problem that I should include as an error message. I tested both R and browser access to the API and it's working again :)
![screenshot from 2018-06-19 09-08-05](https://user-images.githubusercontent.com/10091065/41599252-a26c37e8-73a0-11e8-871a-c81486c48419.png)
![screenshot from 2018-06-19 09-07-13](https://user-images.githubusercontent.com/10091065/41599253-a28b38b4-73a0-11e8-9bb2-5b66df57728f.png)
