### 2018-06-19 #217 @aedobbyn
Hmm I still get the same `parse error: premature EOF` even though that endpoint in the browser now displays data. @cimentadaj are you able to receive data?
