### 2018-06-20 #217 @cimentadaj
Hi @maelle, @aedobbyn and @pachamaltese ! Glad to review this package. I'll submit my review this weekend, but in the mean time, the development version of the package (`devtools::install_github("pachamaltese/oec-r-package")`) seems to work for me:

```
library(oec)
getdata("chl", "chn", 2015)
#> Valid input. Processing 2015 data...
#> # A tibble: 670 x 31
#>     year origin_id destination_id origin_name destination_name id
#>    <dbl> <chr>     <chr>          <chr>       <chr>            <chr>
#>  1  2015 chl       chn            Chile       China            5722
#>  2  2015 chl       chn            Chile       China            5838
#>  3  2015 chl       chn            Chile       China            6251
#>  4  2015 chl       chn            Chile       China            6282
#>  5  2015 chl       chn            Chile       China            6289
#>  6  2015 chl       chn            Chile       China            6352
#>  7  2015 chl       chn            Chile       China            6418
#>  8  2015 chl       chn            Chile       China            6572
#>  9  2015 chl       chn            Chile       China            6577
#> 10  2015 chl       chn            Chile       China            6631
#> # ... with 660 more rows, and 25 more variables: id_len <int>,
#> #   product_name <chr>, group_id <chr>, group_name <chr>,
#> #   export_val <dbl>, export_val_growth_pct <dbl>,
#> #   export_val_growth_pct_5 <dbl>, export_val_growth_val <dbl>,
#> #   export_val_growth_val_5 <dbl>, export_rca <dbl>, import_val <dbl>,
#> #   import_val_growth_pct <dbl>, import_val_growth_pct_5 <dbl>,
#> #   import_val_growth_val <dbl>, import_val_growth_val_5 <dbl>,
#> #   import_rca <dbl>, trade_exchange_val <dbl>, pci <dbl>, pci_rank <dbl>,
#> #   pci_rank_delta <dbl>, top_exporter_code <chr>,
#> #   top_importer_code <chr>, top_importer <chr>, top_exporter <chr>,
#> #   color <chr>

devtools::session_info()
#> Session info -------------------------------------------------------------
#>  setting  value
#>  version  R version 3.4.4 (2018-03-15)
#>  system   x86_64, mingw32
#>  ui       RTerm
#>  language (EN)
#>  collate  English_United States.1252
#>  tz       Europe/Paris
#>  date     2018-06-20
#> Packages -----------------------------------------------------------------
#>  package    * version    date
#>  assertthat   0.2.0      2017-04-11
#>  backports    1.1.2      2017-12-13
#>  base       * 3.4.4      2018-03-15
#>  bindr        0.1.1      2018-03-13
#>  bindrcpp   * 0.2.2      2018-03-29
#>  cli          1.0.0      2017-11-05
#>  compiler     3.4.4      2018-03-15
#>  crayon       1.3.4      2017-09-16
#>  curl         3.2        2018-03-28
#>  datasets   * 3.4.4      2018-03-15
#>  devtools     1.13.5     2018-02-18
#>  digest       0.6.15     2018-01-28
#>  dplyr        0.7.5      2018-05-19
#>  evaluate     0.10.1     2017-06-24
#>  formatR      1.5        2017-04-25
#>  glue         1.2.0.9000 2018-06-19
#>  graphics   * 3.4.4      2018-03-15
#>  grDevices  * 3.4.4      2018-03-15
#>  htmltools    0.3.6      2017-04-28
#>  jsonlite     1.5        2017-06-01
#>  knitr        1.20       2018-02-20
#>  magrittr     1.5        2014-11-22
#>  memoise      1.1.0      2018-03-20
#>  methods    * 3.4.4      2018-03-15
#>  oec        * 2.8.1      2018-06-20
#>  pillar       1.2.2      2018-04-26
#>  pkgconfig    2.0.1      2017-03-21
#>  purrr        0.2.5      2018-05-29
#>  R6           2.2.2      2017-06-17
#>  Rcpp         0.12.17    2018-05-18
#>  rlang        0.2.0.9001 2018-06-20
#>  rmarkdown    1.9        2018-03-01
#>  rprojroot    1.3-2      2018-01-03
#>  stats      * 3.4.4      2018-03-15
#>  stringi      1.1.7      2018-03-12
#>  stringr      1.3.1      2018-05-10
#>  tibble       1.4.2      2018-01-22
#>  tidyselect   0.2.4      2018-02-26
#>  tools        3.4.4      2018-03-15
#>  utf8         1.1.3      2018-01-03
#>  utils      * 3.4.4      2018-03-15
#>  withr        2.1.2      2018-03-18
#>  yaml         2.1.18     2018-03-08
#>  source
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.3)
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.3)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  Github (tidyverse/glue@a2c0f8b)
#>  local
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  Github (hadley/memoise@06d16ec)
#>  local
#>  Github (pachamaltese/oec-r-package@00a367b)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  Github (r-lib/rlang@ba4fb06)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  CRAN (R 3.4.4)
#>  local
#>  CRAN (R 3.4.4)
#>  local
#>  Github (r-lib/withr@79d7b0d)
#>  CRAN (R 3.4.4)
```
