### 2018-06-20 #217 @pachamaltese
Hi @maelle, @aedobbyn and @cimentadaj
I added some error messages to inform the user when the wifi drops or MIT server is down/with problems. Last night I had the same problem (server side) and the browser returned an empty json, then 30 mins later I could obtain the data.
I'll tell Media Lab people about this, but in the meanwhile an error message "patches" server side issues :)
