### 2018-06-20 #217 @aedobbyn
Great, working for me too now!

Just so I know the intent @pachamaltese, you quote variables from the OEC and use `sym` and `!!` to unquote them to avoid the NOTE `no visible binding for global variable {variable}` correct?

A [version](https://github.com/aedobbyn/oec/blob/amanda/R/get_data.R) with the tidyeval stripped out   passes all [tests](https://github.com/aedobbyn/oec/blob/amanda/tests/testthat/test-get_data.R) without warnings so I imagine the NOTE is the only reason.
