### 2018-06-20 #217 @aedobbyn
Totally understand, those NOTEs are annoying :relaxed:

As you probably know, those NOTEs shouldn't stop you from passing a CRAN check, especially if you document them in an . Rbuildignored [`cran-comments.md`](http://r-pkgs.had.co.nz/release.html). Might be cleaner to do that and avoid tidyeval altogether, but I would understand if you want to leave the tidyeval in.
