### 2018-06-20 #217 @aedobbyn
@maelle makes the excellent point that you can let document these global variables in a [`globalVariables.R`](https://github.com/aedobbyn/oec/blob/amanda/R/globalVariables.R) file like this to avoid the `no visible binding for global variable` NOTEs. (Also see, e.g., an example of it in use in her [`opencage`](https://github.com/ropensci/opencage/blob/master/R/globalVariables.R) package.)

 If you `devtools::check()` the `oec` code in that `amanda` branch on my fork you won't get any of those NOTEs and you won't need to use the `sym` + `!!` syntax -- let me know if you want me to shoot you a PR with the code from that fork :)
