### 2018-06-20 #217 @pachamaltese
@aedobbyn thanks! I accept that if and only if you agree to be added as `ctb` because that is a contribution, a good contribution to be precise :)
