### 2018-06-23 #217 @pachamaltese
Whoa. Thanks a lot. This helps to improve the package. "Get_data" was
provided by the other reviewers as an alternative without tidy evaluation.
I'm from cellphone now so I'm eager to implement changes later!

El sáb., jun. 23, 2018 5:59 PM, Jorge Cimentada <notifications@github.com>
escribió:

> Package Review
>
> *Please check off boxes as applicable, and elaborate in comments below.
> Your review is not limited to these topics, as described in the reviewer
> guide*
>
>    - As the reviewer I confirm that there are no conflicts of interest
>    for me to review this work (If you are unsure whether you are in conflict,
>    please speak to your editor *before* starting your review).
>
> Documentation
>
> The package includes all the following forms of documentation:
>
>    - [x ] *A statement of need* clearly stating problems the software is
>    designed to solve and its target audience in README
>    - [x ] *Installation instructions:* for the development version of
>    package and any non-standard dependencies in README
>    - *Vignette(s)* demonstrating major functionality that runs
>    successfully locally
>    - *Function Documentation:* for all exported functions in R help
>    - *Examples* for all exported functions in R Help that run
>    successfully locally
>    - *Community guidelines* including contribution guidelines in the
>    README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and
>    Maintainer (which may be autogenerated via Authors@R).
>
> For packages co-submitting to JOSS
>
>    - The package has an *obvious research application* according to JOSS's
>    definition <http://joss.theoj.org/about#submission_requirements>
>
> The package contains a paper.md matching JOSS's requirements
> <http://joss.theoj.org/about#paper_structure> with:
>
>    - *A short summary* describing the high-level functionality of the
>    software
>    - *Authors:* A list of authors with their affiliations
>    - *A statement of need* clearly stating problems the software is
>    designed to solve and its target audience.
>    - *References:* with DOIs for all those that have one (e.g. papers,
>    datasets, software).
>
> Functionality
>
>    - *Installation:* Installation succeeds as documented.
>    - *Functionality:* Any functional claims of the software been
>    confirmed.
>    - *Performance:* Any performance claims of the software been confirmed.
>    - *Automated tests:* Unit tests cover essential functions of the
>    package
>    and a reasonable range of inputs and conditions. All tests pass on the
>    local machine.
>    - *Packaging guidelines*: The package conforms to the rOpenSci
>    packaging guidelines
>
> Final approval (post-review)
>
>    - *The author has responded to my review and made changes to my
>    satisfaction. I recommend approving this package.*
>
> Estimated hours spent reviewing:
> 1.5 Review Comments
>
> The oec package is well written and definitely fulfills the objective of
> downloading trade date easily into R. I have comments which serve to
> improve the code, the user experience and the readibility of the code. Some
> are more immediate then others so feel free to ignore the small ones or new
> features if time is of essence. However, I would consider them for future
> releases. They are not in any particular order of importance.
>
>    -
>
>    Great opportunity to use a switch statement here
>    <https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L112-L130>
>    .
>    -
>
>    I would use match.args here
>    <https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L62-L64>
>    -
>
>    I'm not sure whether curl retries calling the API if it doesn't
>    respond, but if it doesn't, then you could retry 3-5 times like here
>    <https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L142-L144>.
>    This would be very handy because you call the API several times in other
>    links. Considering that the MIT server has been a bit unstable in the last
>    few days then this could come in handy. I think there's a few packages to
>    do this properly but if you want to avoid adding more dependencies, a
>    function hack is this one using httr:
>
> get_resp <- function(url, attempts_left = 5) {
>
>   stopifnot(attempts_left > 0)
>
>   resp <- httr::GET(url)
>
>   # On a successful GET, return the response
>   if (httr::status_code(resp) == 200) {
>     resp
>   } else if (attempts_left == 1) { # When attempts run out, stop with an error
>     stop("Cannot connect to the MIT API")
>   } else { # Otherwise, sleep a second and try again
>     Sys.sleep(1)
>     get_resp(url, attempts_left - 1)
>   }
> }
>
> Feel free to adapt it to using fromJSON and remove the httr code.
>
>    -
>
>    Maybe I'm a bit ignorant but in this specific case
>    <https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L146-L149>,
>    since you already checked that the user has internet at the beginning of
>    the function, shouldn't this be *only* an MIT problem? Perhaps this
>    error check can be wrapped into a function because it's used 2-3 times.
>    -
>
>    I see that you create several functions called read_from_api_** to
>    create the URL of the API. It would be cleaner and safer to create a single
>    function that is modular to all API urls and you provide different
>    arguments depending on the specific url.
>    -
>
>    Any particular reason why you have get_data and getdata? Seems a bit
>    redundant and not a clear grammatical reason to have two (for example,
>    color and colour in ggplot2). In any case, because they share exactly
>    the same documentation I would suggest documenting them together into a
>    single page using #' @rdname get_data. See
>    http://r-pkgs.had.co.nz/man.html. In fact, because they're the same
>    exact function, if you want to keep both, then I would do something like
>    this:
>
> get_data <- function(all the arguments) {
>  all code here....
> }
> getdata <- function(same arguments) {
> get_data(pass the arguments)
> }
>
> To avoid repetition.
>
>    -
>
>    This is clearly more work, but I think it might be worth it for new
>    users. Instead of relying on your user to search for a country in the 264
>    country names to look for a code, you could create a function that regexps
>    the name of the country and returns the country code. For example,
>    get_countrycode("Chile") or get_countrycode("chile") should yield the
>    same. This can be piped into get_data more intuitively like get_data(get_countrycode("Chile"),
>    get_countrycode("China"), 2010, 2015). Alternatively, the user can
>    search the country tibble to search manually. But note that
>    non-frequent R users will probably search manually and not with regexps.
>    -
>
>    On the same line, you could do something similar for all product
>    tibbles. I would take a similar approach to the way the World Bank R
>    package does it, wbstats. For example
>
> library(wbstats)
> unemploy_vars <- wbsearch(pattern = "unemployment")
> head(unemploy_vars)
>
> Check https://cran.r-project.org/web/packages/wbstats/README.html for
> more examples.
>
> You could have a generic get_product that searches in all product tibbles
> and returns an aggregated tibble with all matches. For example if we
> search for Antiques then we get the same result from oec::hs02 but also a
> new column with indicating that it comes from hs02. This way when you
> have similar matches then the user can disambiguate where it comes from.
>
> Of course this is a lot of work, but here's a very quick and crude
> approach:
>
> get_products <- function(name) {
>   # Grab all the names of all hs datasets
>   all_datastr <-
>     stringr::str_subset(
>       data(package = "oec")$results[, "Item"],
>       "^hs"
>     )
>
>   # get the datasets, create the type_product column, bind them all together
>   # and do the search
>   all_datastr %>%
>     purrr::map(get) %>%
>     purrr::map2_dfr(all_datastr, ~ {.x$type_product <- .y; .x}) %>%
>     dplyr::filter(stringr::str_detect(product_name, name))
> }
>
> get_products("Gold")
> get_products("Animals")
>
> Misc:
>
> In the oec website and Github repo the installation code chunk still has
> the old repo (devtools::install_github("pachamaltese/oec-r-package"). In
> fact, try doing an overall repo search and update all URL's. The
> DESCRIPTION file needs to be updated in the BugReports field as well.
>
> I like that you're very defensive in what you accept in the function but
> try to help the user in pointing out where something went wront. For
> example, here
> <https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L47-L51> I
> will paste what the user wrote and when appropriate (not here because
> there's too many countries), list the available options the user could've
> chosen.
>
> Great package!
>
