### 2018-06-25 #217 @aedobbyn
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

2

First off, I'm really excited for this package and the data it's going to make available to people in tidy format.

I think the modifications you've put in place so far and the ones suggested by @cimentadaj are all very useful. (Many of those I was going to suggest myself, but I won't repeat Jorge!) +1 especially to trying the server multiple times (you might even implement exponential backoff) and making things a bit more concise with `match.args` and `switch`.

A few other notes:

- An alternative to @cimentadaj's suggestion to allow users to `get_countrycode` using a regex match would be to allow users to supply either the `country_code` *or* the `country` in `get_data`; so `get_data("chl", "chn", 2015, "sitc")` would return the same data as `get_data("Chile", "China", 2015, "sitc")`. While this approach gives the user less leeway in the input they can supply than a regex would, I think most of the countries in `country_codes$country` are represented the way most people would expect, which saves some work on your end.

- It looks like you [allow the user to supply "all" as origin and destination](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L179), but this isn't reflected in the documentation. Beyond simply mentioning that this is an option, it may even be worth setting `"all"` as the default value for `origin` and `destination` so that the user can get up and running with the function without looking through country codes.

I'll note that supplying `"all"` for both `origin` and `destination` does fail at the moment because during the [join trade flows step](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L278) you end up with `top_importer.x`, `top_importer.y`, etc. so operations expecting a single `top_importer` error out. So this case will need a bit of special treatment.

```
library(oec)
getdata("all", "all", 2015)
#> Valid input. Processing 2015 data...
#> Error in .f(.x[[i]], ...): object 'top_importer' not found
```

- The [docs say](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L10) that `classification` defaults to `"sitc"`, but that [doesn't seem to be the case](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L41) at the moment. I'm all in favor of more defaults for the function, so users can get up and running more quickly!

- There are some quick wins to increase [test coverage](https://codecov.io/gh/pachamaltese/oec/src/6da195837d28fb9638f89e466fd9a33816dca6f8/R/get_data.R) that would make the package more stable in the long term if the API data availability changes. I'd add tests for all the different classification types; that way if something happens to one of them down the road, your test will fail and Travis will alert you. Currently there is only a test for `classification = "sitc"`.

- I'd also suggest using `webmockr` to in the tests as well, as [@maelle mentioned](https://github.com/ropensci/onboarding/issues/217#issuecomment-398150814)

- The styling generally looks good but I'd recommend a `[styler](https://github.com/r-lib/styler)::style_pkg()` just to catch the remaining small things.

- I would consider changing `oec_r_package.Rproj` to `oec.Rproj` for consistency (and I second @cimentadaj's point that it's good to do an entire sweep of the package to make sure that there are no more instances of `oec_r_package` or `oec-r-package` left).

- I believe only the .Rmd needs to go in the `vignettes` directory. I like that you show in a vignette how you went about tidying the data that the user finds in the package.

- I would add one or two examples of `get_data` to the readme or the pkgdown site so that people don't have to find the vignette or go to the docs to get a preview of the package. (Makes it more likely for people to download the package instead of clicking away, imo).

- You may already be in the process of doing the full switchover from `getdata` to `get_data` but just to make sure -- anywhere where `getdata` was previously used e.g. in the vignettes, examples etc. should be replaced with `get_data` and of course the `getdata` function and its test file can be deleted.

- You might consider documenting what each of these fields returned from the API mean. If there isn't a good link to point people to in the `@return`, you could include a tibble  in the package that the user can access with something like `variable_definitions`.

I think that's it from me! Looking forward to seeing this package get some good use!
