### 2018-06-25 #217 @maelle
Many thanks for your reviews @aedobbyn and @cimentadaj, they seem very thorough and useful! 😺

I learnt about the global variables "trick" myself [here](https://github.com/ropensci/onboarding/issues/36#issuecomment-215123733) 😉
