### 2018-06-25 #217 @maelle
By the way you can add reviewers as `ctb` but also as `rev`  cf https://ropensci.github.io/dev_guide/building.html#authorship
