### 2018-06-25 #217 @maelle
Of course! As per [our guide for authors](https://ropensci.github.io/dev_guide/onboarding-guide-for-authors.html) "we ask that you respond to reviewers’ comments within 2 weeks of the last-submitted review, but you may make updates to your package or respond at any time.". 😺
