### 2018-07-07 #217 @pachamaltese
@maelle can I have more time to implement all the suggestions? I'm working on them and the coolness shown on the feedback requires time to implement :)
