### 2018-07-14 #217 @pachamaltese
@maelle @aedobbyn @cimentadaj

Thanks a lot for the feedback. Here's a list of todos after the feedback, I shall be working on this after the commit I just made as a few parts require dedication. Now I finally have time to organize this and produce a good result. Have a good weekend!!

- [x] Use a switch statement [here](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L112-L130)
- [ ] Use match.args [here](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L62-L64)
- [ ] Re-try API calls [here](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L142-L144) by adapting this function:
```r
get_resp <- function(url, attempts_left = 5) {

  stopifnot(attempts_left > 0)

  resp <- httr::GET(url)

  # On a successful GET, return the response
  if (httr::status_code(resp) == 200) {
    resp
  } else if (attempts_left == 1) { # When attempts run out, stop with an error
    stop("Cannot connect to the MIT API")
  } else { # Otherwise, sleep a second and try again
    Sys.sleep(1)
    get_resp(url, attempts_left - 1)
  }
}
```
- [ ] Write a function to report server side connection errors
- [x] Write general `read_from_api` function
- [ ] Include `get_data` suggestions in `getdata`
- [ ] To keep or not to keep tidyeval
- [ ] Create a function that regexps the name of the country and returns the country code. For example, `get_countrycode("Chile")` or `get_countrycode("chile")` should yield the same. This can be piped into `get_data` more intuitively like `get_data(get_countrycode("Chile")`
- On the same line, do something similar for all product tibbles. For example
```r
library(wbstats)
unemploy_vars <- wbsearch(pattern = "unemployment")
head(unemploy_vars)
```
- [ ]  Create generic `get_product` that searches in all product tibbles and returns an aggregated tibble with all matches. For example if we search for Antiques then we get the same result from `oec::hs02` but also a new column with indicating that it comes from `hs02`. Very quick and crude approach:
```r
get_products <- function(name) {
  # Grab all the names of all hs datasets
  all_datastr <-
    stringr::str_subset(
      data(package = "oec")$results[, "Item"],
      "^hs"
    )

  # get the datasets, create the type_product column, bind them all together
  # and do the search
  all_datastr %>%
    purrr::map(get) %>%
    purrr::map2_dfr(all_datastr, ~ {.x$type_product <- .y; .x}) %>%
    dplyr::filter(stringr::str_detect(product_name, name))
}

get_products("Gold")
get_products("Animals")
```
- [ ] An alternative to @cimentadaj's suggestion to allow users to get_countrycode using a regex match would be to allow users to supply either the `country_code` or the country in `get_data`; so `get_data("chl", "chn", 2015, "sitc")` would return the same data as `get_data("Chile", "China", 2015, "sitc")`. While this approach gives the user less leeway in the input they can supply than a regex would, I think most of the countries in country_codes$country are represented the way most people would expect, which saves some work on your end.
- [x] Include "all" as origin and destination in documentation and as default value for origin and destination so that the user can get up and running with the function without looking through country codes.
- [ ] Fix join trade flows step to deal with `top_importer.x, top_importer.y` when using "all" as argument
- [x] Move `sitc` default to arguments instead of "if missing"
- [ ] Use `webmockr` in tests
- [ ] Use Styler (https://github.com/r-lib/styler) instead of Lint
- [x] Rename `oec_r_package.Rproj` to `oec.Rproj`
- [ ] Check for non-Rmd files in `vignettes/`
- [ ] Add one or two examples of `get_data` to the readme or the pkgdown site so that people don't have to find the vignette or go to the docs to get a preview of the package.
- [ ] Switch `getdata` to `get_data`
- [ ] Document results from API and explain a bit about RCA, monetary units, etc.
