### 2018-07-29 #217 @pachamaltese
@maelle I' m almost ready!! I implemented all of the suggestions except webmocker and styler, I'm doing that tonight. Thanks a lot @aedobbyn @cimentadaj, thanks to your comments the package improved a ton and now you can jsut write `get_data("CHILE", "ArGeNtiNa") and the internals will make it work :)
