### 2018-07-29 #217 @pachamaltese
@maelle now I only have webmockr implementation in unit tests and explaining a bit about result from API :D
