### 2018-07-30 #217 @maelle
@pachamaltese thanks for the update! Note that @sckott's `webmockr` doesn't support `httr` yet, only `crul`, so you can either switch to `crul`, or keep using `httr` and use https://github.com/nealrichardson/httptest
