### 2018-07-30 #217 @sckott
thanks @maelle - httr integration is operational in a branch on github. install like `remotes::install_github("ropensci/vcr@httr-integration")` and use it just like the crul integration. Would love any feedback. It should go to CRAN soon as we don't require any changes in httr itself
