### 2018-07-30 #217 @pachamaltese
thanks @maelle @sckott !
@maelle now I completed all the changes except for the webmockr part which I am not sure how to implement
