### 2018-07-30 #217 @maelle
Ok thanks for the update @pachamaltese!

@sckott is there any guide, or any example using `webmockr` with `httr` in the wild (a toy example?)? I only know of https://ropensci.github.io/http-testing-book/
