### 2018-08-02 #217 @sckott
👋 @pachamaltese So I know exactly what you want to do, do you want to do mocking with `webmockr` where you simply match requests, but don't cache the responses? Or do you want more what `vcr` does where it matches requests and caches responses?  `vcr` is a good fit for using in your test suite, whereas `webmockr` can be used for tests, but doesn't have to be, and you do have to cache your own responses with `webmockr` if you want them cached. Let me know then I can give examples
