### 2018-08-02 #217 @pachamaltese
Hi @sckott. I think it would be better to use `vcr`, match request and cache responses, specially when sometimes the server acts weird. Thanks a lot!!
