### 2018-08-02 #217 @sckott
Okay. `httr` integration is not yet on CRAN. You need github versions of `vcr` and `webmockr`. Install like `remotes::install_github("ropensci/vcr@httr-integration")`

The usage is no different from if you were using `crul`. So any examples, etc. in the docs should suffice. I just made a tiny package to demo using `httr` and `vcr` together https://github.com/sckott/catfact so you can see what the setup is like.

more in depth docs for vcr/webmockr here https://ropensci.github.io/http-testing-book/
