### 2018-08-22 #217 @pachamaltese
Yes! Thanks a lot for the useful comments.
I just kept tidyeval for consistency.

Best,

On Wed, Aug 22, 2018, 1:32 AM Maëlle Salmon <notifications@github.com>
wrote:

> @pachamaltese <https://github.com/pachamaltese> cool, congrats! So you're
> done with all comments?
>
