### 2018-08-22 #217 @maelle
@aedobbyn @cimentadaj can you please have a last look at the package? Thanks!
