### 2018-08-24 #217 @aedobbyn
Looking good @pachamaltese!

Here are a few notes. I hope to be able to to look through more thoroughly this weekend but in case I'm not able to I wanted to give you something. By topic, in no particular order:

**`get_data`**

- In `get_data`, the warnings `No encoding supplied: defaulting to UTF-8.` occur during the [`parse()` call](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L144)
  - This warning isn't very useful to the user (since they can't do anything about it) and there could potentially be a lot of them since you have one call per year. I'd either set the encoding explicitly using `parse(encoding = "UTF-8")` or suppress the message

- In [`read_from_api`](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L114) you have an `attempts_left` argument but there's no way for the user to change this. Not required but you could think about including this as a user-defined variable in `get_data`

- This is super nitpicky but the [processing data message](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L110) is a little misleading because it seems like a progress bar when really it messages all the years upfront. That means that if there were a problem with a certain year these messages wouldn't help you pin down which year failed. Maybe something like `sprintf("\nProcessing %s to %s data...", range(years)[1], range(years)[2])` would be better. Otherwise, you could include the message of which year we're on in the `read_from_api` function.

**`get_countrycode`**

- The error message for country code that doesn't exist could be moe informative than `length(countrycode) != 0 is not TRUE`

- You might want to include a mapping from "us" to "united states", "uk" to "united kingdom", and other similar situations since I can see people trying `get_countrycode("us")` and being surprised when they don't get a US result

**Other**

- In the devtools install line, it should be `devtools::install_github("pachamaltese/oec")`

- You shouldn't need to prefix package functions with `oec::` (e.g. https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L87), I think

- I'm not sure what the correct protocol for the `fixtures` dir . Looks like you use it in `tests/testthat/helper-oec.R` to get `vcr` configurations. The presence of the extra directory does give you a NOTE on `devtools::check()`, so I think it should at least be buildignored

- Should update the vignette titles (line 7 for both). On `devtools::release_checks()`: `WARNING: placeholder 'Vignette Title' detected in 'title' field and/or 'VignetteIndexEntry' for: oec-data.Rmd,oec.Rmd`

- May want to update the year on your LICENSE and in the README to 2018

- What was the rationale for keeping the tidyeval instead of removing it and using a `globals.R` file?

This is looking really solid overall -- will be excited to use it very soon!
