### 2018-08-24 #217 @pachamaltese
WHOA !! thanks a lot @aedobbyn !!!!!!!!!
I'm making changes this weekend. For now, the rationale for keeping the tidyeval was just to mimic some design decisions I like, for example those in the `highcharter` package by @jbkunst.
