### 2018-08-26 #217 @aedobbyn
Thanks for implementing those changes so quickly! Last bit of suggestions from me, I think:

- In your `get_data` `@return` documentation I would point people to your handy [reference](http://pacha.hk/oec/articles/oec.html) about what each of the columns means since a lot of these are non-obvious.

- If you feel it's worth it (though I don't think strictly necessary), you could refactor out some of the redundancies in the `get_data` function. For instance, you use the same [`rename` + `mutate` flow](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L179) in [each](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L219) [extraction](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L238), so this could be pulled out into a utility function. Similarly, the [checks](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L58) at the beginning of the function making sure that the years specified are in bounds for the `classification` could be stored in a lookup table. This way you only need one of these in `get_data` instead of five.

- Right now you only allow years up to 2016, but in the coming years will the OEC have data past that year? Instead of updating this by hand is there a way to check what the greatest year available is, maybe on package load and store that as a variable?

- I think in all cases you can replace single `&`s with `&&`s and `|`s with `||`s

- Small typo [here](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L10) ("lower of equal to")

- [Unnecessary line](https://github.com/pachamaltese/oec/tree/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L97), I think

- Future improvement for `get_productcode` would be a `word_boundary` boolean param for an exact word match. That would take care of the first result for `get_productcode("wine")` being "Live swine" 😆

- Very small detail -- you might want to put the `top_exporter_code` and `top_exporter` cols next to each other in the output since they reference the same thing. Right now they're sometimes separated by a few columns (e.g. `trade_exchange_val`). Same for `top_importer_code` and `top_importer`

- Should re-run pkgdown once new changes are implemented
