### 2018-08-26 #217 @pachamaltese
*thanks a lot* @aedobbyn !!
I still have some pendings that I am adding to this new list
btw, after all of this cool changes you suggested, can I contact you by email? it's for a related project that benefited a lot from this feedback
