### 2018-08-26 #217 @cimentadaj
Great job @pachamaltese. I think @aedobbyn has covered most of the major problems. Here are some minor concerns.

* https://github.com/pachamaltese/oec/blob/master/R/get_countrycode.R#L51-L52 should say 'vector' rather than package.

* Perhaps you can move [this](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L112-L161) outside the main function so it's cleaner.

* In general, I would create tests that check that the **content** of the request resembles what you requested and not only the format of the tibble. I would add a few tests checking whether `get_data` returned the requested country as well as for the different years. Maybe this applies to some topics as well.  I just had a nasty bug in one of my packages related to this and it was awful!

* [Here](https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L145) I would check whether it has a 'try-error' to be safe. This makes sense because you're using try anyways.

Hope these are useful!
