### 2018-09-05 #217 @pachamaltese
@aedobbyn @cimentadaj thanks a lot!!
I had hectic days at the office, here are some changes that I'll try to complete ASAP

- [x] In your get_data @return documentation I would point people to your handy reference about what each of the columns means since a lot of these are non-obvious.
- [x] *If you feel it's worth it (though I don't think strictly necessary), you could refactor out some of the redundancies in the get_data function. For instance, you use the same rename + mutate flow in each extraction, so this could be pulled out into a utility function. Similarly, the checks at the beginning of the function making sure that the years specified are in bounds for the classification could be stored in a lookup table. This way you only need one of these in get_data instead of five.* ok , I added a wrapper to tidy IDs, probably more wrappers coming during the week
- [x] *Right now you only allow years up to 2016, but in the coming years will the OEC have data past that year? Instead of updating this by hand is there a way to check what the greatest year available is, maybe on package load and store that as a variable?* there's no way with the current API without loading "trash" with the year value (for future work, sure!)
- [ ] I think in all cases you can replace single &s with &&s and |s with ||s
- [x] Small typo here ("lower of equal to")
- *Unnecessary line, I think (https://github.com/pachamaltese/oec/blob/73f4a7493aa649311f88f60dcbd0cfd4a203c895/R/get_data.R#L97)* Without it the checks fail
- [ ] *Future improvement for get_productcode would be a word_boundary boolean param for an exact word match. That would take care of the first result for get_productcode("wine") being "Live swine" laughing* Added to issues https://github.com/pachamaltese/oec/issues/22
- [x] *Very small detail -- you might want to put the top_exporter_code and top_exporter cols next to each other in the output since they reference the same thing. Right now they're sometimes separated by a few columns (e.g. trade_exchange_val). Same for top_importer_code and top_importer* noted, thanks!! I moved that and others columns to a more "logical" order
- [x] Should re-run pkgdown once new changes are implemented
- [x] *https://github.com/pachamaltese/oec/blob/master/R/get_countrycode.R#L51-L52 should say 'vector' rather than package.* thanks!! totally re-phrased now
