### 2018-09-12 #217 @pachamaltese
@aedobbyn @cimentadaj  Hi. Thanks for the feedback! there are four point I can't solve without breaking the check :S
This weekend I' ll try to solve https://github.com/pachamaltese/oec/blob/master/R/get_data.R#L112-L161 as it is totally cleaner with that change that without
