### 2018-09-27 #217 @maelle
:wave: @pachamaltese did you get any chance to make progress on the package? When you do, please summarize your response again (using your checklist again I guess :wink:).
