### 2018-09-27 #217 @pachamaltese
@maelle Hi, I was not able to test a lot because the server has been down a lot actually.
I updated the list to reflect the only 3 point I could not fix. Those can be more tricky actually!
