### 2018-10-16 #217 @pachamaltese
Hi Maelle
The server is just acting weird.
Can we put this on hold?
Yesterday I spoke to Amanda and Jorge about providing an alternative API
that would be hassle-free.
Best

On Tue, Oct 16, 2018, 1:25 AM Maëlle Salmon <notifications@github.com>
wrote:

> @pachamaltese <https://github.com/pachamaltese> any update?
>
