### 2018-10-16 #217 @pachamaltese
Exactly, at the present time that' the problem.
Last year it was different

On Tue, Oct 16, 2018, 11:56 AM Maëlle Salmon <notifications@github.com>
wrote:

> Do you mean you don't trust the data provider enough with stability?
>
