### 2018-12-06 #217 @pachamaltese
Hi Maelle

Sorry about the long silence!
I have a tentative PostgreSQL database and now I only have to work on the
API. The good part is that both Jorge and Amanda are collaborating on this.
If you also want to participate please do not hesitate to give ideas to
create a better API.

Best!

—————

*Mauricio Vargas Sepúlveda 帕夏*
Do you like Data Science? visit pacha.hk
你爱科学数据专吗？你走pacha.hk

El jue., 6 de dic. de 2018 a la(s) 07:00, Maëlle Salmon (
notifications@github.com) escribió:

> @pachamaltese <https://github.com/pachamaltese> 👋, any news?
>
