### 2018-12-06 #217 @pachamaltese
sure !! I have created this organization to make my last 3 years of work
fully reproducible
https://github.com/tradestatistics

—————

*Mauricio Vargas Sepúlveda 帕夏*
Do you like Data Science? visit pacha.hk
你爱科学数据专吗？你走pacha.hk

El jue., 6 de dic. de 2018 a la(s) 10:22, Maëlle Salmon (
notifications@github.com) escribió:

> Cool to read! Is there a public repo we can link from here in case someone
> reads the thread and gets curious?
>
