### 2019-01-02 #217 @pachamaltese
@maelle Hi!!
Sorry for the very long silence. After all the issues with the "server is down" and whatever, I ended up creating my own project.

A new package based on `oec` is now available here https://github.com/tradestatistics/r-package

But in broader terms, tradestatistics.io is something new I made with a help from the awesome community and everything is pure R (even the API, which is where the new `tradestatistics` package takes the data)

There are tons of changes such as the ability to inspect different tables, filter by commodity code length, etc provided I made my own PSQL DB.

Probably I am using a laser beam to open a tuna can, but anyways I've tried do to something to foster reproducibility.

Do you have ideas to rename the package? `tradestatistics` sounds ok, but maybe `tsapi` is more appropiate.

Least but not last, api.tradestatistics.io/friend is a (boring?) joke I made with the API. It should be easy to complete the joke with the correct query.
