### 2019-01-03 #217 @maelle
:wave: @pachamaltese! Happy New Year! Congrats on the project (and fun endpoint)!

Should we close this issue? And when your new package is ready you can open a [pre-submission inquiry](https://github.com/ropensci/software-review/issues/new?template=B-submit-a-presubmission-inquiry.md) (if it does the exact same things at the package using the broken server used to do, it should be in scope, but better asking first).

I find `tradestatistics` a better name because `tsapi` can be misleading since ts is often an abbreviation for time series.
