### 2019-01-03 #217 @maelle
> I'll prepare the new submission during the next hour

Maybe best open a pre-submission inquiry and wait until the package is more mature until a full submission (unless the package _is_ mature already).
